Split Provider::Terraform in multiple logical block
